+++ src/App.js
@@ -7,8 +7,8 @@ import "./App.css"
 const AppContainer=styled.div`
 bottom: 0px;
 margin:0px;
-width:100vw;
-height:100%;
+max-width:100vw;
+min-height:100%;
 background-image: url(${spaceship});
 background-repeat: no-repeat;
 background-size:cover;
@@ -20,7 +20,7 @@ background-size:cover;
 function App() {
   return (
     <AppContainer>
-        <Router/>
+       <Router/>
     </AppContainer>
        
   );
