+++ src/pages/AdminList/AdmListPage.js
@@ -1,35 +1,10 @@
 import React from 'react'
 import { useHistory } from 'react-router-dom';
-import styled from 'styled-components'
-import { useRequestData } from "../hooks/useRequestData";
-import { useProtectedPage } from "../hooks/useProtectPage";
-import AdmHeader from './AdmHeader';
+import { useRequestData } from "../../hooks/useRequestData";
+import { useProtectedPage } from "../../hooks/useProtectPage";
+import AdmHeader from '../../components/AdmHeader';
+import {ContainerList, Card, ButtonApplication} from "./styled"
 
-const ContainerList= styled.div`
-   display:grid;
-  grid-template-columns: 1fr 1fr 1fr 1fr;
-  width:100vw;
- `
-const Card = styled.div`
- width:200px;
- margin:1em;
- border:solid 2px;
- padding:1em;
- background-color: rgba(0,0,0,0.5);
- p,h4{
-   color:white;
- }
- 
- button{
-   width:200px;
-   padding:0.5em;
-   background-color:gray;
-   font-size: 20px;
-   bottom:0px;
-   margin: 30px 0px;
-   
- }
-`
 function AdmListPage() {
 
     useProtectedPage();
@@ -54,7 +29,7 @@ function AdmListPage() {
          <img src={`https://picsum.photos/200/200?a=${i}]`}/>
           <h4 key={trip.id}>{ trip.name}</h4>
           <p>planeta: {trip.planet}</p>
-          <button onClick={()=>{goToDetail(trip.id)}}> Ver detalhes</button>
+          <ButtonApplication onClick={()=>{goToDetail(trip.id)}}> Ver detalhes</ButtonApplication>
       </ Card>
       );
     })
