+++ src/pages/Home/styled.js
@@ -1,0 +1,19 @@
+import styled from 'styled-components'
+
+export const HomeContainer=styled.div`
+ display:flex;
+ justify-content:center;
+ height:100vh; 
+`
+export const TextContainer=styled.div`
+  width: 900px;
+  height: 30%;
+  margin: 2em;
+  padding: 2em;
+  text-align:justify;
+  border:solid 2px;
+  background-color: rgba(0,0,0,0.5);
+  p{
+    color:white;
+  }
+`
