+++ src/pages/ListTrips/ListTripsPage.js
@@ -1,35 +1,10 @@
 import React, { useEffect } from 'react'
 import { useHistory } from 'react-router-dom';
-import styled from 'styled-components'
-import { useRequestData } from "../hooks/useRequestData";
-import Header from './Header'
+import { useRequestData } from "../../hooks/useRequestData";
+import {ContainerList, Card, ButtonApplication} from './styled'
+import Header from '../../components/Header'
 
-const ContainerList= styled.div`
-  display:grid;
-  grid-template-columns: 1fr 1fr 1fr 1fr;
-  width:100vw;
- `
 
-const Card = styled.div`
- width:200px;
- margin:1em;
- border:solid 2px;
- padding:1em;
- background-color: rgba(0,0,0,0.5);
- p,h4{
-   color:white;
- }
- 
- button{
-   width:200px;
-   padding:0.5em;
-   background-color:gray;
-   font-size: 20px;
-   bottom:0px;
-   margin: 30px 0px;
-   
- }
-`
 
 function ListTripsPage() {
   
@@ -38,7 +13,7 @@ function ListTripsPage() {
     
   );
 
-  console.log(getSpaceTrip)
+  
   const history=useHistory()
 
   const goToApplication=(id)=>{
@@ -55,7 +30,7 @@ function ListTripsPage() {
           <p>planeta: {trip.planet}</p>
           <p>duração: {trip.durationInDays}</p>
           <p>Data: {trip.date}</p>
-          <button onClick={()=>{goToApplication(trip.id)}}>Quero ir</button>
+          <ButtonApplication onClick={()=>{goToApplication(trip.id)}}>Quero ir</ButtonApplication>
       </ Card>
       );
     })
