+++ src/pages/Login/LoginPage.js
@@ -1,35 +1,10 @@
 import React, { useEffect, useState } from "react";
 import { useHistory } from 'react-router-dom';
 import axios from "axios";
-import styled from 'styled-components'
-import Header from './Header'
+import Header from '../../components/Header'
+import {ContainerForm,Input,ButtonForm } from "./styled"
 
 
-const ContainerForm= styled.div`
-  display:flex;
-  flex-direction: column;
-  align-items: center;
-  h1{
-    color:white;
-    margin-top:80px;
-  }
-  height:100vh;
- `
-const Input= styled.input`
-  padding: 10px;
-  width: 300px;
-  margin: 0.5em;
-`
-const ButtonForm=styled.button`
-background-color: rgba(0,0,0,0.8);
-height:70px;
-margin-top:1.5em;
-color:white;
-padding:0.5em;
-font-family: 'Audiowide', cursive;
-border-radius:30px;   
-font-size: 20px;
-`
 function LoginPage() {
   const [admEmail, setAdmEmail] = useState("");
   const [admPassword, setAdmPassword] = useState("");
