+++ src/pages/TripDetail/TripDetailsPage.js
@@ -1,49 +1,17 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
-import { useProtectedPage } from "../hooks/useProtectPage";
-import styled from 'styled-components'
+import { useProtectedPage } from "../../hooks/useProtectPage";
+import{DetailContainer,Card, Detail, Photo, ButtonAcept} from "./styled"
 import { useParams, useHistory } from 'react-router-dom'
-import AdmHeader from "./AdmHeader";
+import AdmHeader from "../../components/AdmHeader";
 
-const DetailContainer = styled.div`
- display:flex;
- justify-content:space-around;
- height:100vh;
- `
-const Card = styled.div`
- width:500px;
- border:solid 2px;
- padding:1em;
- margin-top:1em;
- background-color: rgba(10,23,55,0.5);
- p,h4,h2{
-   color:white;
- }
- `
- const Detail= styled.div`
-    display:flex;
-    justify-content:space-around;
-    margin:1em;
- `
- const Photo = styled.img`
-    width:50px;
-    height:50px;
-    border-radius:50%;
-    margin-right:1em;
-  `
-  const ButtonAcept = styled.button`
-    width:50px;
-    padding:0.5em;
-    background-color:black;
-    color:white;
-    margin: 10px 10px;
-    border-radius:30px;   
-  `
+
 function TripDetailsPage() {
 
   const { id } = useParams()
   const [trips, setTrips] = useState({});
   const [candidates, setCandidates]= useState([]);
+  const [approved, setApproved]= useState([]);
   
   useProtectedPage();
 
@@ -64,8 +32,10 @@ function TripDetailsPage() {
         .then((res) => {
         setTrips(res.data.trip);
         setCandidates(res.data.trip.candidates)
+        setApproved(res.data.trip.approved)
         console.log( "Detalhes",res)
         console.log("trips",trips)
+        console.log("Approved",approved)
       })
       .catch((err) => {
         console.log(err);
@@ -91,8 +61,8 @@ function TripDetailsPage() {
           }else{
             alert("Ahhhhh, tente novamente!")
           }
-          
-            console.log (response)
+           getTripDetail()
+            console.log ("Aprovados",response.data)
         })
         .catch(e => {
             console.log(e)
@@ -105,13 +75,31 @@ function TripDetailsPage() {
     <div>
       <Detail>
        <Photo src={`https://picsum.photos/200/200?a=${i}]`}/>
-        <p key={candidate.id}>Olá meu nome é {candidate.name}, tenho {candidate.age} anos, vivo no {candidate.country} e sou {candidate.profession}</p>
+        <p key={candidate.id}>Olá meu nome é {candidate.name}, tenho {candidate.age} anos, meu país é {candidate.country} e sou {candidate.profession}</p>
       </Detail>
       <p>{candidate.applicationText}</p>
       <ButtonAcept onClick={()=>{aceptCandidate(candidate.id,true)}}>Yes</ButtonAcept>
       <ButtonAcept onClick={()=>{aceptCandidate(candidate.id,false)}}>No</ButtonAcept>
     </div>
-    
+    );
+  })
+
+  const aprovados= approved.map((candidate,i)=>{
+    return(
+    <div>
+      <Detail key={candidate.id}>
+      <hr/>
+       <Photo src={`https://picsum.photos/200/200?a=${i}]`}/>
+        <div>
+          <p><strong>Nome: </strong>{candidate.name}</p> 
+          <p><strong>Idade: </strong>{candidate.age}</p> 
+          <p><strong>País: </strong>{candidate.country}</p> 
+          <p><strong>Profissão: </strong>{candidate.profession}</p> 
+        </div>
+        <hr/>
+      </Detail>
+     
+    </div>
     );
   })
 
@@ -126,12 +114,19 @@ function TripDetailsPage() {
           <p>Data: {trips.date}</p>
           <p>Duração:{trips.durationInDays}</p>
           <p>Descrição: {trips.description}</p>
+          <br/>
+          <hr/>
+          <br/>
+          <h2>Candidatos aprovados</h2>
+          {aprovados}
         </Card>
 
         <Card>
           <h2>Controle de viajantes</h2>
           <h4>Atividade: {trips.name}</h4>
           {candidatos}
+          
+
         </Card>
       </DetailContainer>
   </div>
